fixed control registers

+++ kernel/include/cputables.h
@@ -99,49 +99,49 @@ typedef union
     struct
     {
         /** @brief Carry Flag */
-        uint32_t CF : 1;
+        uint64_t CF : 1;
         /** @brief Reserved */
-        uint32_t always_one : 1;
+        uint64_t always_one : 1;
         /** @brief Parity Flag */
-        uint32_t PF : 1;
+        uint64_t PF : 1;
         /** @brief Reserved */
-        uint32_t _reserved0 : 1;
+        uint64_t _reserved0 : 1;
         /** @brief Auxiliary Carry Flag */
-        uint32_t AF : 1;
+        uint64_t AF : 1;
         /** @brief Reserved */
-        uint32_t _reserved1 : 1;
+        uint64_t _reserved1 : 1;
         /** @brief Zero Flag */
-        uint32_t ZF : 1;
+        uint64_t ZF : 1;
         /** @brief Sign Flag */
-        uint32_t SF : 1;
+        uint64_t SF : 1;
         /** @brief Trap Flag */
-        uint32_t TF : 1;
+        uint64_t TF : 1;
         /** @brief Interrupt Enable Flag */
-        uint32_t IF : 1;
+        uint64_t IF : 1;
         /** @brief Direction Flag */
-        uint32_t DF : 1;
+        uint64_t DF : 1;
         /** @brief Overflow Flag */
-        uint32_t OF : 1;
+        uint64_t OF : 1;
         /** @brief I/O Privilege Level */
-        uint32_t IOPL : 2;
+        uint64_t IOPL : 2;
         /** @brief Nested Task */
-        uint32_t NT : 1;
+        uint64_t NT : 1;
         /** @brief Reserved */
-        uint32_t _reserved2 : 1;
+        uint64_t _reserved2 : 1;
         /** @brief Resume Flag */
-        uint32_t RF : 1;
+        uint64_t RF : 1;
         /** @brief Virtual 8086 Mode */
-        uint32_t VM : 1;
+        uint64_t VM : 1;
         /** @brief Alignment Check */
-        uint32_t AC : 1;
+        uint64_t AC : 1;
         /** @brief Virtual Interrupt Flag */
-        uint32_t VIF : 1;
+        uint64_t VIF : 1;
         /** @brief Virtual Interrupt Pending */
-        uint32_t VIP : 1;
+        uint64_t VIP : 1;
         /** @brief ID Flag */
-        uint32_t ID : 1;
+        uint64_t ID : 1;
         /** @brief Reserved */
-        uint32_t _reserved3 : 10;
+        uint64_t _reserved3 : 10;
     };
     uint64_t raw;
 } RFLAGS;
@@ -151,33 +151,33 @@ typedef union CR0
     struct
     {
         /** @brief Protection Enable */
-        uint32_t PE : 1;
+        uint64_t PE : 1;
         /** @brief Monitor Coprocessor */
-        uint32_t MP : 1;
+        uint64_t MP : 1;
         /** @brief Emulation */
-        uint32_t EM : 1;
+        uint64_t EM : 1;
         /** @brief Task Switched */
-        uint32_t TS : 1;
+        uint64_t TS : 1;
         /** @brief Extension Type */
-        uint32_t ET : 1;
+        uint64_t ET : 1;
         /** @brief Numeric Error */
-        uint32_t NE : 1;
+        uint64_t NE : 1;
         /** @brief Reserved */
-        uint32_t _reserved0 : 10;
+        uint64_t _reserved0 : 10;
         /** @brief Write Protect */
-        uint32_t WP : 1;
+        uint64_t WP : 1;
         /** @brief Reserved */
-        uint32_t _reserved1 : 1;
+        uint64_t _reserved1 : 1;
         /** @brief Alignment Mask */
-        uint32_t AM : 1;
+        uint64_t AM : 1;
         /** @brief Reserved */
-        uint32_t _reserved2 : 10;
+        uint64_t _reserved2 : 10;
         /** @brief Mot Write-through */
-        uint32_t NW : 1;
+        uint64_t NW : 1;
         /** @brief Cache Disable */
-        uint32_t CD : 1;
+        uint64_t CD : 1;
         /** @brief Paging */
-        uint32_t PG : 1;
+        uint64_t PG : 1;
     };
     uint64_t raw;
 } CR0;
@@ -197,9 +197,9 @@ typedef union CR3
     struct
     {
         /** @brief Not used if bit 17 of CR4 is 1 */
-        uint32_t PWT : 1;
+        uint64_t PWT : 1;
         /** @brief Not used if bit 17 of CR4 is 1 */
-        uint32_t PCD : 1;
+        uint64_t PCD : 1;
         /** @brief Base of PML4T/PML5T */
         uint64_t PDBR;
     };
@@ -211,53 +211,53 @@ typedef union CR4
     struct
     {
         /** @brief Virtual-8086 Mode Extensions */
-        uint32_t VME : 1;
+        uint64_t VME : 1;
         /** @brief Protected-Mode Virtual Interrupts */
-        uint32_t PVI : 1;
+        uint64_t PVI : 1;
         /** @brief Time Stamp Disable */
-        uint32_t TSD : 1;
+        uint64_t TSD : 1;
         /** @brief Debugging Extensions */
-        uint32_t DE : 1;
+        uint64_t DE : 1;
         /** @brief Page Size Extensions */
-        uint32_t PSE : 1;
+        uint64_t PSE : 1;
         /** @brief Physical Address Extension */
-        uint32_t PAE : 1;
+        uint64_t PAE : 1;
         /** @brief Machine Check Enable */
-        uint32_t MCE : 1;
+        uint64_t MCE : 1;
         /** @brief Page Global Enable */
-        uint32_t PGE : 1;
+        uint64_t PGE : 1;
         /** @brief Performance Monitoring Counter */
-        uint32_t PCE : 1;
+        uint64_t PCE : 1;
         /** @brief Operating System Support */
-        uint32_t OSFXSR : 1;
+        uint64_t OSFXSR : 1;
         /** @brief Operating System Support */
-        uint32_t OSXMMEXCPT : 1;
+        uint64_t OSXMMEXCPT : 1;
         /** @brief User-Mode Instruction Prevention */
-        uint32_t UMIP : 1;
+        uint64_t UMIP : 1;
         /** @brief Linear Address 57bit */
-        uint32_t LA57 : 1;
+        uint64_t LA57 : 1;
         /** @brief VMX Enable */
-        uint32_t VMXE : 1;
+        uint64_t VMXE : 1;
         /** @brief SMX Enable */
-        uint32_t SMXE : 1;
+        uint64_t SMXE : 1;
         /** @brief Reserved */
-        uint32_t _reserved0 : 1;
+        uint64_t _reserved0 : 1;
         /** @brief FSGSBASE Enable */
-        uint32_t FSGSBASE : 1;
+        uint64_t FSGSBASE : 1;
         /** @brief PCID Enable */
-        uint32_t PCIDE : 1;
+        uint64_t PCIDE : 1;
         /** @brief XSAVE and Processor Extended States Enable */
-        uint32_t OSXSAVE : 1;
+        uint64_t OSXSAVE : 1;
         /** @brief Reserved */
-        uint32_t _reserved1 : 1;
+        uint64_t _reserved1 : 1;
         /** @brief SMEP Enable */
-        uint32_t SMEP : 1;
+        uint64_t SMEP : 1;
         /** @brief SMAP Enable */
-        uint32_t SMAP : 1;
+        uint64_t SMAP : 1;
         /** @brief Protection-Key Enable */
-        uint32_t PKE : 1;
+        uint64_t PKE : 1;
         /** @brief Reserved */
-        uint32_t _reserved2 : 9;
+        uint64_t _reserved2 : 9;
     };
     uint64_t raw;
 } CR4;
@@ -267,7 +267,7 @@ typedef union CR8
     struct
     {
         /** @brief Task Priority Level */
-        uint32_t TPL : 1;
+        uint64_t TPL : 1;
     };
     uint64_t raw;
 } CR8;
@@ -277,27 +277,27 @@ typedef union PageFaultErrorCode
     struct
     {
         /** @brief When set, the page fault was caused by a page-protection violation. When not set, it was caused by a non-present page. */
-        uint32_t P : 1;
+        uint64_t P : 1;
         /** @brief 	When set, the page fault was caused by a write access. When not set, it was caused by a read access. */
-        uint32_t W : 1;
+        uint64_t W : 1;
         /** @brief When set, the page fault was caused while CPL = 3. This does not necessarily mean that the page fault was a privilege violation. */
-        uint32_t U : 1;
+        uint64_t U : 1;
         /** @brief When set, one or more page directory entries contain reserved bits which are set to 1. This only applies when the PSE or PAE flags in CR4 are set to 1. */
-        uint32_t R : 1;
+        uint64_t R : 1;
         /** @brief When set, the page fault was caused by an instruction fetch. This only applies when the No-Execute bit is supported and enabled. */
-        uint32_t I : 1;
+        uint64_t I : 1;
         /** @brief When set, the page fault was caused by a protection-key violation. The PKRU register (for user-mode accesses) or PKRS MSR (for supervisor-mode accesses) specifies the protection key rights. */
-        uint32_t PK : 1;
+        uint64_t PK : 1;
         /** @brief When set, the page fault was caused by a shadow stack access. */
-        uint32_t SS : 1;
+        uint64_t SS : 1;
         /** @brief Reserved */
-        uint32_t _reserved0 : 8;
+        uint64_t _reserved0 : 8;
         /** @brief When set, the fault was due to an SGX violation. The fault is unrelated to ordinary paging. */
-        uint32_t SGX : 1;
+        uint64_t SGX : 1;
         /** @brief Reserved */
-        uint32_t _reserved1 : 16;
+        uint64_t _reserved1 : 16;
     };
-    uint32_t raw;
+    uint64_t raw;
 } PageFaultErrorCode;
 
 // ! TODO: UNTESTED!
@@ -306,18 +306,18 @@ typedef union SelectorErrorCode
     struct
     {
         /** @brief When set, the exception originated externally to the processor. */
-        uint32_t External : 1;
+        uint64_t External : 1;
         /** @brief IDT/GDT/LDT Table
          *  @details 0b00 - The Selector Index references a descriptor in the GDT.
          *  @details 0b01 - The Selector Index references a descriptor in the IDT.
          *  @details 0b10 - The Selector Index references a descriptor in the LDT.
          *  @details 0b11 - The Selector Index references a descriptor in the IDT.
          */
-        uint32_t Table : 2;
+        uint64_t Table : 2;
         /** @brief The index in the GDT, IDT or LDT. */
-        uint32_t Idx : 13;
+        uint64_t Idx : 13;
         /** @brief Reserved */
-        uint32_t Reserved : 16;
+        uint64_t Reserved : 16;
     };
     uint64_t raw;
 } SelectorErrorCode;
